Break up TeleConsole interface into separate files
